base_be/store/api/v1/product/serializer.py
from rest_framework import serializers
from store.models.product import Product


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ["id", "title", "description"]
