base_be/requirements.txt
Django==4.2.6
django-mptt
autopep8
jedi
pylint
pillow
boto3
django-storages
python-dotenv
djangorestframework
markdown
django-filter
